Changed promo and review section
